+++ src/ch/epfl/leb/sass/simulator/Simulator.java
@@ -19,6 +19,8 @@
  */
 package ch.epfl.leb.sass.simulator;
 
+import ch.epfl.leb.sass.utils.images.ImageS;
+import ch.epfl.leb.sass.utils.images.internal.DefaultImageS;
 import ij.ImageStack;
 import ij.process.ImageProcessor;
 import java.io.File;
@@ -48,7 +50,7 @@ public interface Simulator {
      * Generates a new image and adds it to the internal stack.
      * @return newly generated image
      */
-    public ImageProcessor getNextImage();
+    public ImageS getNextImage();
     
     /**
      * Increments the simulation by one time step without creating an image.
@@ -96,9 +98,10 @@ public interface Simulator {
     
     /**
      * Returns internal stack with all generated images.
+     * 
      * @return internal stack
      */
-    public ImageStack getStack();
+    public ImageS getStack();
     
     /**
      *